Move parse_args
